test: 로그인 관련 스토리북 수정
